fix(test): rename _personas.ts -> personas.ts (kebab-case CI gate)
Phase 2B Naming Standards CI rejects underscores in source filenames.
The leading underscore was a bashism for "non-test helper" — Jest already
ignores it because the filename does not end in .test.ts, so the rename
has zero functional impact.
